[AURON #1358] Rename package name for hadoop-shim module
