+++ spark-extension/src/main/java/org/apache/spark/sql/auron/JniBridge.java
@@ -26,10 +26,10 @@ import org.apache.hadoop.fs.Path;
 import org.apache.spark.SparkEnv;
 import org.apache.spark.TaskContext;
 import org.apache.spark.TaskContext$;
-import org.apache.spark.auron.FSDataInputWrapper;
-import org.apache.spark.auron.FSDataInputWrapper$;
-import org.apache.spark.auron.FSDataOutputWrapper;
-import org.apache.spark.auron.FSDataOutputWrapper$;
+import org.apache.auron.hadoop.fs.FSDataInputWrapper;
+import org.apache.auron.hadoop.fs.FSDataInputWrapper$;
+import org.apache.auron.hadoop.fs.FSDataOutputWrapper;
+import org.apache.auron.hadoop.fs.FSDataOutputWrapper$;
 import org.apache.spark.sql.auron.memory.OnHeapSpillManager;
 import org.apache.spark.sql.auron.memory.OnHeapSpillManager$;
 import org.apache.spark.sql.auron.util.TaskContextHelper$;
